Broke out BookmarkInspiration component from profile.js to BookmarkInspiration.js in profile folder
